ref(perf): Move performance utility to its own folder
The number of utility functions for the performance product has steadily been
increasing. This moves many of them into their own dedicated folder at
`app/utils/performance` to make for better organization.
