+++ src/sentry/static/sentry/app/utils/measurements/index.tsx
@@ -1,5 +1,5 @@
 import {getDuration} from 'app/utils/formatters';
-import {Vital} from 'app/views/performance/transactionVitals/types';
+import {Vital} from 'app/utils/performance/vitals/types';
 
 export function formattedValue(record: Vital | undefined, value: number): string {
   if (record && record.type === 'duration') {
