+++ src/sentry/static/sentry/app/utils/performance/quickTrace/utils.ts
@@ -1,0 +1,5 @@
+import {Event, EventTransaction} from 'app/types/event';
+
+export function isTransaction(event: Event): event is EventTransaction {
+  return event.type === 'transaction';
+}
