+++ src/sentry/static/sentry/app/utils/performance/trends/trendsDiscoverQuery.tsx
