+++ src/sentry/static/sentry/app/utils/performance/vitals/vitalsCardsDiscoverQuery.tsx
