+++ src/sentry/static/sentry/app/views/performance/transactionDetails/content.tsx
@@ -22,6 +22,10 @@ import {t} from 'app/locale';
 import {Organization, Project} from 'app/types';
 import {Event, EventTag} from 'app/types/event';
 import {trackAnalyticsEvent} from 'app/utils/analytics';
+import * as QuickTraceContext from 'app/utils/performance/quickTrace/quickTraceContext';
+import QuickTraceQuery, {
+  QuickTraceQueryChildrenProps,
+} from 'app/utils/performance/quickTrace/quickTraceQuery';
 import Projects from 'app/utils/projects';
 import {appendTagCondition, decodeScalar} from 'app/utils/queryString';
 import Breadcrumb from 'app/views/performance/breadcrumb';
@@ -30,8 +34,6 @@ import {transactionSummaryRouteWithQuery} from '../transactionSummary/utils';
 import {getTransactionDetailsUrl} from '../utils';
 
 import EventMetas from './eventMetas';
-import * as QuickTraceContext from './quickTraceContext';
-import QuickTraceQuery, {QuickTraceQueryChildrenProps} from './quickTraceQuery';
 
 type Props = {
   organization: Organization;
