+++ src/sentry/static/sentry/app/views/performance/transactionSummary/latencyChart.tsx
@@ -12,12 +12,11 @@ import {t} from 'app/locale';
 import {OrganizationSummary} from 'app/types';
 import {trackAnalyticsEvent} from 'app/utils/analytics';
 import EventView from 'app/utils/discover/eventView';
+import HistogramQuery from 'app/utils/performance/histogram/histogramQuery';
+import {HistogramData} from 'app/utils/performance/histogram/types';
+import {computeBuckets, formatHistogramData} from 'app/utils/performance/histogram/utils';
 import {decodeScalar} from 'app/utils/queryString';
 import theme from 'app/utils/theme';
-
-import {computeBuckets, formatHistogramData} from '../charts/utils';
-import HistogramQuery from '../transactionVitals/histogramQuery';
-import {HistogramData} from '../transactionVitals/types';
 
 const NUM_BUCKETS = 50;
 const QUERY_KEYS = [
@@ -74,7 +73,7 @@ class LatencyChart extends React.Component<Props, State> {
     this.setState({zoomError: true});
   };
 
-  bucketWidth(data: HistogramData[]) {
+  bucketWidth(data: HistogramData) {
     // We can assume that all buckets are of equal width, use the first two
     // buckets to get the width. The value of each histogram function indicates
     // the beginning of the bucket.
@@ -94,7 +93,7 @@ class LatencyChart extends React.Component<Props, State> {
     );
   }
 
-  renderChart(data: HistogramData[]) {
+  renderChart(data: HistogramData) {
     const {location} = this.props;
     const {zoomError} = this.state;
