+++ src/sentry/static/sentry/app/views/performance/transactionVitals/constants.tsx
@@ -1,99 +1,14 @@
 import {t} from 'app/locale';
 import {SelectValue} from 'app/types';
-import {measurementType, WebVital} from 'app/utils/discover/fields';
+import {WebVital} from 'app/utils/discover/fields';
+import {DataFilter} from 'app/utils/performance/histogram/types';
+import {WEB_VITAL_DETAILS} from 'app/utils/performance/vitals/constants';
+import {VitalGroup} from 'app/utils/performance/vitals/types';
 import theme from 'app/utils/theme';
-
-import {DataFilter, Vital, VitalGroup} from './types';
 
 export const NUM_BUCKETS = 100;
 
 export const PERCENTILE = 0.75;
-
-export const WEB_VITAL_DETAILS: Record<WebVital, Vital> = {
-  [WebVital.FP]: {
-    slug: 'fp',
-    name: t('First Paint'),
-    acronym: 'FP',
-    description: t(
-      'Render time of the first pixel loaded in the viewport (may overlap with FCP).'
-    ),
-    failureThreshold: 3000,
-    type: measurementType(WebVital.FP),
-  },
-  [WebVital.FCP]: {
-    slug: 'fcp',
-    name: t('First Contentful Paint'),
-    acronym: 'FCP',
-    description: t(
-      'Render time of the first image, text or other DOM node in the viewport.'
-    ),
-    failureThreshold: 3000,
-    type: measurementType(WebVital.FCP),
-  },
-  [WebVital.LCP]: {
-    slug: 'lcp',
-    name: t('Largest Contentful Paint'),
-    acronym: 'LCP',
-    description: t(
-      'Render time of the largest image, text or other DOM node in the viewport.'
-    ),
-    failureThreshold: 4000,
-    type: measurementType(WebVital.LCP),
-  },
-  [WebVital.FID]: {
-    slug: 'fid',
-    name: t('First Input Delay'),
-    acronym: 'FID',
-    description: t(
-      'Response time of the browser to a user interaction (clicking, tapping, etc).'
-    ),
-    failureThreshold: 300,
-    type: measurementType(WebVital.FID),
-  },
-  [WebVital.CLS]: {
-    slug: 'cls',
-    name: t('Cumulative Layout Shift'),
-    acronym: 'CLS',
-    description: t(
-      'Sum of layout shift scores that measure the visual stability of the page.'
-    ),
-    failureThreshold: 0.25,
-    type: measurementType(WebVital.CLS),
-  },
-  [WebVital.TTFB]: {
-    slug: 'ttfb',
-    name: t('Time to First Byte'),
-    acronym: 'TTFB',
-    description: t(
-      "The time that it takes for a user's browser to receive the first byte of page content."
-    ),
-    failureThreshold: 600,
-    type: measurementType(WebVital.TTFB),
-  },
-  [WebVital.RequestTime]: {
-    slug: 'ttfb.requesttime',
-    name: t('Request Time'),
-    acronym: 'RT',
-    description: t(
-      'Captures the time spent making the request and receiving the first byte of the response.'
-    ),
-    failureThreshold: 600,
-    type: measurementType(WebVital.RequestTime),
-  },
-};
-
-// translate known short form names into their long forms
-export const LONG_WEB_VITAL_NAMES = Object.fromEntries(
-  Object.values(WEB_VITAL_DETAILS).map(value => {
-    return [value.slug, value.name];
-  })
-);
-
-export const WEB_VITAL_ACRONYMS = Object.fromEntries(
-  Object.values(WEB_VITAL_DETAILS).map(value => {
-    return [value.slug, value.acronym];
-  })
-);
 
 export const FILTER_OPTIONS: SelectValue<DataFilter>[] = [
   {label: t('Exclude Outliers'), value: 'exclude_outliers'},
