+++ src/sentry/static/sentry/app/views/performance/transactionVitals/vitalCard.tsx
@@ -28,11 +28,13 @@ import {
   getDuration,
 } from 'app/utils/formatters';
 import getDynamicText from 'app/utils/getDynamicText';
+import {HistogramData} from 'app/utils/performance/histogram/types';
+import {computeBuckets, formatHistogramData} from 'app/utils/performance/histogram/utils';
+import {Vital} from 'app/utils/performance/vitals/types';
+import {VitalData} from 'app/utils/performance/vitals/vitalsCardsDiscoverQuery';
 import {Theme} from 'app/utils/theme';
 import {stringifyQueryObject, tokenizeSearch} from 'app/utils/tokenizeSearch';
-import {VitalData} from 'app/views/performance/vitalDetail/vitalsCardsDiscoverQuery';
 
-import {computeBuckets, formatHistogramData} from '../charts/utils';
 import {VitalBar} from '../landing/vitalsCards';
 import {
   VitalState,
@@ -43,7 +45,7 @@ import {
 
 import {NUM_BUCKETS, PERCENTILE} from './constants';
 import {Card, CardSectionHeading, CardSummary, Description, StatNumber} from './styles';
-import {HistogramData, Rectangle, Vital} from './types';
+import {Rectangle} from './types';
 import {asPixelRect, findNearestBucketIndex, getRefRect, mapPoint} from './utils';
 
 type Props = {
@@ -55,7 +57,7 @@ type Props = {
   vital: WebVital;
   vitalDetails: Vital;
   summaryData: VitalData | null;
-  chartData: HistogramData[];
+  chartData: HistogramData;
   colors: [string];
   eventView: EventView;
   min?: number;
@@ -160,7 +162,7 @@ class VitalCard extends React.Component<Props, State> {
   renderSummary() {
     const {vitalDetails: vital, colors, eventView, organization, min, max} = this.props;
     const summary = this.summary;
-    const {slug, name, description, failureThreshold} = vital;
+    const {slug, name, description, poorThreshold} = vital;
 
     const column = `measurements.${slug}`;
 
@@ -196,7 +198,7 @@ class VitalCard extends React.Component<Props, State> {
         <SummaryHeading>
           <CardSectionHeading>{`${name} (${slug.toUpperCase()})`}</CardSectionHeading>
           {summary === null || this.showVitalColours() ? null : summary <
-            failureThreshold ? (
+            poorThreshold ? (
             <Tag>{t('Pass')}</Tag>
           ) : (
             <StyledTag>{t('Fail')}</StyledTag>
@@ -477,12 +479,12 @@ class VitalCard extends React.Component<Props, State> {
   getFailureSeries() {
     const {theme, chartData, vitalDetails: vital} = this.props;
     const failureRate = this.failureRate;
-    const {failureThreshold, type} = vital;
+    const {poorThreshold, type} = vital;
     if (this.state.refDataRect === null || this.state.refPixelRect === null) {
       return null;
     }
 
-    let failureBucket = findNearestBucketIndex(chartData, failureThreshold);
+    let failureBucket = findNearestBucketIndex(chartData, poorThreshold);
     if (failureBucket === null) {
       return null;
     }
@@ -570,8 +572,8 @@ class VitalCard extends React.Component<Props, State> {
           t(
             'Fails threshold at %s.',
             type === 'duration'
-              ? getDuration(failureThreshold / 1000, 2, true)
-              : formatFloat(failureThreshold, 2)
+              ? getDuration(poorThreshold / 1000, 2, true)
+              : formatFloat(poorThreshold, 2)
           ),
           '</strong>',
           '</span>',
